% Auto-generated: do not edit by hand
\name{dedEditableDiv}

\alias{dedEditableDiv}

\title{EditableDiv component}

\description{
EditableDiv is div component with editable content storing its innerHTML in the state.
}

\usage{
dedEditableDiv(children=NULL, id=NULL, value=NULL, className=NULL,
style=NULL)
}

\arguments{
\item{children}{Named list. The chilren components}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Character. The div innerHTML.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{style}{Named list. Defines CSS styles which will override styles previously set.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

